requirements.txt
numpy==1.21.5
tensorflow==2.11.0
waymo-open-dataset-tf-2-11-0==1.6.0
tqdm==4.66.1
